requirements.txt
numpy==1.21.5
opencv-python==4.5.5.64
matplotlib==3.5.2
seaborn==0.11.2
scikit-learn==1.0.2
albumentations==1.3.0
ultralytics==8.0.20
PyYAML==6.0
jupyter==1.0.0